requirements.txt
flask
openai
python-dotenv
requests
langdetect
reportlab
sqlalchemy
